// src/marker/enums/marker.enum.ts
export enum MarkerType {
  POOR_LIGHTING = 'POOR_LIGHTING',
  SUSPECTED_DRUG_TRAFFICKING = 'SUSPECTED_DRUG_TRAFFICKING',
  HARASSMENT_REPORTS = 'HARASSMENT_REPORTS',
  UNSAFE_AREA = 'UNSAFE_AREA',
  ISOLATED_LOCATION = 'ISOLATED_LOCATION',
  HIGH_CRIME_RATE = 'HIGH_CRIME_RATE',
  CATCALLING_ZONE = 'CATCALLING_ZONE',
  STALKING_REPORTS = 'STALKING_REPORTS',

  SAFE_SPOT = 'SAFE_SPOT',
  POLICE_STATION = 'POLICE_STATION',
  WOMEN_SUPPORT_CENTER = 'WOMEN_SUPPORT_CENTER',
  WELL_LIT_AREA = 'WELL_LIT_AREA',
  SECURITY_CAMERA = 'SECURITY_CAMERA',
  EMERGENCY_BUTTON = 'EMERGENCY_BUTTON',
  WOMEN_ONLY_SPACE = 'WOMEN_ONLY_SPACE',
  TRUSTED_ESTABLISHMENT = 'TRUSTED_ESTABLISHMENT',

  UNSAFE_BUS_STOP = 'UNSAFE_BUS_STOP',
  SAFE_TAXI_POINT = 'SAFE_TAXI_POINT',
  BIKE_SHARING_SAFE = 'BIKE_SHARING_SAFE',
  METRO_SAFE_EXIT = 'METRO_SAFE_EXIT',

  NIGHT_DANGER_ZONE = 'NIGHT_DANGER_ZONE',
  WEEKEND_RISK_AREA = 'WEEKEND_RISK_AREA',
  CONSTRUCTION_HAZARD = 'CONSTRUCTION_HAZARD',
  BROKEN_INFRASTRUCTURE = 'BROKEN_INFRASTRUCTURE',
}